Display in Swagger-UI the version of OpenId4VP supported.
