+++ src/main/resources/application.properties
@@ -7,6 +7,7 @@ server.error.includeStacktrace=ALWAYS
 server.port=8080
 spring.webflux.static-path-pattern=/public/**
 spring.webflux.webjars-path-pattern=/webjars/**
+spring.messages.basename=i18n/messages
 
 #
 # Verifier options
