+++ src/main/resources/i18n/messages.properties
@@ -1,0 +1,1 @@
+openid4vp-version=OpenId4VP {0}
