+++ src/main/resources/public/css/style.css
@@ -1,0 +1,3 @@
+.topbar {
+    display: none;
+}
